Improve grass and grass_block rendering
- Create nomip render layer for disabling mipmap on certain textures.
- Disable mipmapping on grass, tallgrass, and sugar cane.
- Randomly rotate and shift uvs for grass_block top texture to break up
  the grid pattern.
- Determine render layer during asset load time.
- Improve mipmap filter by not accumulating completely transparent
  samples.

+++ src/nbt.cpp
@@ -290,7 +290,7 @@ bool ParseTag(RingBuffer& rb, Tag& tag, MemoryArena& arena) {
     tag.tag = long_array_tag;
   } break;
   default: {
-    fprintf(stderr, "Unknown NBT type: %d\n", tag.type);
+    fprintf(stderr, "Unknown NBT type: %d\n", (int)tag.type);
     return 0;
   }
   }
